Shame on me.  A lot of cleanup and fixes.

+++ aux/prep.sh
@@ -1,19 +1,21 @@
 
 
-pkill -f /opt/snag/bin/snagc
-ps -ef |grep 'snagc.pl' |grep -v grep | awk '{print $2}' | xargs kill 2>/dev/null
-
-export PP_INCLUDES=/var/tmp/snag/script/includes/pp_includes
-export SNAGC_INCLUDES=/var/tmp/snag/script/includes/snagc_includes
-export SNAGW_INCLUDES=/var/tmp/snag/script/includes/snagw_includes
-export SNAGX_INCLUDES=/var/tmp/snag-x/script/includes/snagx_includes
+if [[ -z $SNAG_BUILD_DIR ]]
+then
+  echo "please . {SNAG_BUILD_DIR}/snagrc prior to execution"
+  exit
+fi
 
 echo "PP_INCLUDES=$PP_INCLUDES"
 echo "SNAGC_INCLUDES=$SNAGC_INCLUDES"
 echo "SNAGX_INCLUDES=$SNAGX_INCLUDES"
 echo "SNAGW_INCLUDES=$SNAGW_INCLUDES"
 
-cd /var/tmp/snag \
+
+pkill -f /opt/snag/bin/snagc
+ps -ef |grep 'snagc.pl' |grep -v grep | awk '{print $2}' | xargs kill 2>/dev/null
+
+cd $SNAG_BUILD_DIR \
 && perl Makefile.PL && make && make test && make install \
 && cd /var/tmp \
 && perl /opt/snag/perls/current/bin/snagc.pl --compile --debug \
@@ -21,8 +23,4 @@ cd /var/tmp/snag \
 && perl /opt/snag/perls/current/bin/snagp.pl --compile --debug \
 && perl /opt/snag/perls/current/bin/snagx.pl --compile --debug \
 cd -
-
-#&& mv snagc /opt/snag/bin/  \
-#&& tar zvcpf snag.5.12.1.tar.gz  snag/snag.conf  snag/bin/snagc snag/log snag/conf snag/sbin \
-#&& mv snag.5.12.1.tar.gz /var/tmp/
